+++ src/computational_graph/common.jl
@@ -1,53 +1,3 @@
-import ..Filter
-# import ..Wirreducible  #remove all polarization subdiagrams
-# import ..Girreducible  #remove all self-energy inseration
-import ..NoHartree
-import ..NoFock
-import ..NoBubble  # true to remove all bubble subdiagram
-import ..Proper  #ver4, ver3, and polarization diagrams may require to be irreducible along the transfer momentum/frequency
-
-# import ..DiagType
-# import ..Vacuum
-# import ..Tadpole
-# import ..FermiPropagator
-# import ..BosePropagator
-# import ..FermiSelfEnergy
-# import ..BoseSelfEnergy
-# import ..VertexDiag
-# import ..GncDiag
-# import ..GndDiag
-
-import ..Composite
-import ..ChargeCharge
-import ..SpinSpin
-import ..UpUp
-import ..UpDown
-import ..Response
-
-import ..Instant
-import ..Dynamic
-import ..D_Instant
-import ..D_Dynamic
-import ..AnalyticProperty
-
-import ..symbol
-import ..short
-
-import ..Interaction
-
-import ..DiagramPara
-
-import ..innerTauNum
-
-# unique id
-# uid() = abs(rand(Int)) % 10000
-
-# let z = 0
-#     global function uid()
-#         z += 1
-#         return z
-#     end
-# end
 const _counter = [0,]
 
 mutable struct DType
@@ -81,10 +31,3 @@ end
 function labelreset()
     _labelcounter[1] = 0
 end
-
-function getK(loopNum::Int, loopIdx::Int)
-    k = zeros(loopNum)
-    k[loopIdx] = 1.0
-    return k
-end
-
